requirements.txt
Django==3.2.20
djangorestframework==3.10.3
pillow==9.5.0
frezee
xhtml2pdf
django-ckeditor